requirements.txt
bitcoin-utils==0.5.9
python-dotenv==1.0.0
redis==4.5.4
celery==5.2.7